--- src/main/java/model/dao/ProfileDao.java
@@ -1,5 +1,0 @@
-package model.dao;
-
-public class ProfileDao {
-
-}
